Use std::vector as internal storage for CPDF_Array
Replace the usage of CFX_ArrayTemplate inside CPDF_Array, which
has non-standard APIs such as GetSize() returns int.

R=tsepez@chromium.org

Review URL: https://codereview.chromium.org/1867183002 .

+++ core/fpdfapi/fpdf_edit/fpdf_edit_create.cpp
@@ -96,7 +96,7 @@ int32_t PDF_CreatorAppendObject(const CPDF_Object* pObj,
       }
       offset += 1;
       const CPDF_Array* p = pObj->AsArray();
-      for (uint32_t i = 0; i < p->GetCount(); i++) {
+      for (size_t i = 0; i < p->GetCount(); i++) {
         CPDF_Object* pElement = p->GetObjectAt(i);
         if (pElement->GetObjNum()) {
           if (pFile->AppendString(" ") < 0) {
@@ -1199,7 +1199,7 @@ int32_t CPDF_Creator::WriteDirectObj(uint32_t objnum,
       }
       m_Offset += 1;
       const CPDF_Array* p = pObj->AsArray();
-      for (uint32_t i = 0; i < p->GetCount(); i++) {
+      for (size_t i = 0; i < p->GetCount(); i++) {
         CPDF_Object* pElement = p->GetObjectAt(i);
         if (pElement->GetObjNum()) {
           if (m_File.AppendString(" ") < 0) {
